판 떼기 V2

+++ src/main/java/com/example/panttegi/list/dto/ListResponseDto.java
@@ -4,12 +4,14 @@ import com.example.panttegi.common.BaseDtoDataType;
 import com.example.panttegi.list.entity.BoardList;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class ListResponseDto implements BaseDtoDataType {
 
     private final Long id;
     private final String title;
-    private final Double position;
+    private final BigDecimal position;
 
     public ListResponseDto(BoardList boardList) {
         this.id = boardList.getId();
